Update README.md so it's a bit more offensive
Once I was contacted by a recruiter who had checked out one of my public gitbub repos.
Never again, my README.md now slighly frowns upon that. 
I want this to be a platform to collaborate with other awesome coders, not a platform for recruitment.
Whoever is reading this might feel differently, and that's totally fine. 
In 10 years, I'm probably going to think this was one of the most cringiest, stupidest things I have written.
Ok then, but it's cooler to write it in a github commit than to write it on facebook or linkedin.